server/requirements.txt
Flask >=0.12
twilio~=6.0.0
Pillow
requests
numpy